requirements.txt
baml-py==0.213.0
python-dotenv==1.0.0
kuzu>=0.7.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy
ipywidgets>=8.0.0
node2vec>=0.4.5
networkx>=3.0
anthropic[vertex]>=0.68.0
google-auth>=2.0.0
google-cloud-aiplatform>=1.40.0
streamlit>=1.28.0
